+++ CHANGELOG.md
@@ -1,3 +1,24 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.5]
 
-* TODO: Describe initial release.
+* NEW: Add ability to set the x and y offset of scroll view
+* NEW: Add getter methods for x and y scroll values
+* NEW: Add getter methods for overscrolled widget width and height
+* NEW: Add getter methods for container that holds the overscrolled widget width and height
+
+## [0.0.4]
+
+* NEW: Add optional listener for x and y scroll values
+* NEW: Add velocityFactor to make fling velocity configurable
+* FIX: Prevent fling and gesture animation mixing up
+
+## [0.0.3]
+
+* NEW: Bring project files to root level
+
+## [0.0.2]
+
+* NEW: Add Flutter mention to project
+
+## [0.0.1]
+
+* NEW: Initial Release.